normo_backend/pyproject.toml
[project]
name = "normo-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "mohammad.alnajdawi", email = "mohammad.alnajdawi@teamviewer.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "ipykernel>=6.30.1",
    "ipython>=9.4.0",
    "langchain>=1.1.0",
    "langchain-community>=0.3.27",
    "langchain-openai>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langgraph>=0.6.4",
    "openai>=1.99.6",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "pydantic-settings>=2.6.1",
    "pypdf>=4.0.0",
    "langchain-text-splitters>=0.3.0",
    "motor>=3.3.2",
    "pymongo>=4.6.0",
    "docling>=2.21.1",
    "transformers>=4.30.0",
    "pillow>=10.0.0",
]

[project.scripts]
normo-backend = "normo_backend.main:main"
normo-cli = "normo_backend.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
